update run and email

+++ source/run/acquisition.sh
@@ -3,23 +3,59 @@ printf "\n\n===== Recompiling acquisition.o =====\n\n"
 cd ../Acquisition
 make clear
 make acquisition.o
-cp acquisition.o ../run/libs
+cp acquisition.o ../run/exec
+
 
 
 # Run acquisition
 printf "\n\n===== Running acquisition =====\n\n"
 cd ../run
 
-NOW=$( date +%s ) # Get time epoch to use as folder name
-FOLDER=../data/$NOW
 
-mkdir $FOLDER
 
-libs/acquisition.o $FOLDER # |& tee ../data/output/output.txt
+# Run an infinite acquisition
+while true; do
 
-zip -r "../data/$NOW.zip" $FOLDER
+    # Wait for a few seconds
+    sleep 5
 
-rm -rf $FOLDER
+    # Get time to use as folder name
+    NOW_AS_DATE=$( date +"%Y%m%d_%H%M%S" ) 
+    FOLDER=../data/$NOW_AS_DATE
+
+    mkdir $FOLDER
+
+    exec/acquisition.o $FOLDER |& tee ../data/output/output.txt
+
+    statusAcquisition=0
+
+
+    if [ $? -ne 0 ]; then
+
+        ((statusAcquisition++))
+
+        if [ $statusAcquisition -eq 3 ]; then
+
+            printf "\n\n\nACQUISITION FAILED!!! Number of retries: $statusAcquisition times.\n\n\n"
+
+            rm -rf $FOLDER
+
+            break
+        fi
+
+        rm -rf $FOLDER
+
+        continue
+    fi
+
+
+    zip -r "../data/$NOW_AS_DATE.zip" $FOLDER
+
+    rm -rf $FOLDER
+done
+
+printf "\n\n\nFinishing acquisition.\n\n"
+
 
 
 # SEND EMAIL
